PHP Captcha Generator
This project is based on PHP captcha generatation. A combination of two
programs is used to achieve this. The image developed by generate.php is
used as captcha image. It creates the captcha image based on the random
number created while visiting captcha.php. A random number in the range
1000-9999 is created for this purpose. To make it more secure to use,
numerous lines are added over the image generated. This image will then
displayed at captcha.php . Each time you refresh the page a new number
is created.